backend/requirements.txt
# FastAPI framework and server
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
pydantic-settings>=2.8.1
python-multipart>=0.0.6
email-validator>=2.0.0

# Database connections
azure-cosmos>=4.3.1
motor>=3.1.2  # MongoDB async driver (optional backup)

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.24.0
authlib>=1.1.0  # For OAuth2 and OpenID Connect
PyJWT>=2.6.0  # For JWT handling
cryptography>=39.0.1  # For secure token generation


# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# API documentation
fastapi-versioning>=0.10.0

# Utilities
Pillow>=9.5.0  # For image processing
python-barcode>=0.14.0  # For barcode generation/processing
python-dateutil>=2.8.2  # For date handling
requests>=2.28.2  # For external API calls

# Deployment and production
gunicorn>=20.1.0
